Rename sequencer package and update imports
