--- ValueSequencer.py/__init__.py
@@ -1,21 +1,0 @@
-from __future__ import annotations
-
-from .value_sequencer import CValueSequencer
-from .bit_value_sequencer import CBitValueSequencer
-from .line_value_sequencer import CLineValueSequencer
-from .trigram_value_sequencer import CTrigramValueSequencer
-from .hexagram_value_sequencer import CHexagramValueSequencer
-from .chexagram_sequences import CHexagramSequences, CHexagram
-from .sequences import Sequences, ValueType
-
-__all__ = [
-    "CValueSequencer",
-    "CBitValueSequencer",
-    "CLineValueSequencer",
-    "CTrigramValueSequencer",
-    "CHexagramValueSequencer",
-    "CHexagramSequences",
-    "CHexagram",
-    "Sequences",
-    "ValueType",
-]
