+++ value_sequencer_py/hexagram_value_sequencer.py
@@ -2,9 +2,9 @@ from __future__ import annotations
 
 from typing import List, Optional
 
-from .trigram_value_sequencer import CTrigramValueSequencer
-from .value_sequencer import CValueSequencer
-from .sequences import Sequences
+from value_sequencer_py.trigram_value_sequencer import CTrigramValueSequencer
+from value_sequencer_py.value_sequencer import CValueSequencer
+from value_sequencer_py.sequences import Sequences
 
 
 class CHexagramValueSequencer(CValueSequencer):
